Split 2014 and 2015 logic test functionality review logic into 2 classes
OCD-2532
