+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/listing/Edition2014LegacyListingValidator.java
@@ -12,12 +12,12 @@ import gov.healthit.chpl.validation.listing.reviewer.CertificationStatusReviewer
 import gov.healthit.chpl.validation.listing.reviewer.DeveloperStatusReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.FieldLengthReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.Reviewer;
-import gov.healthit.chpl.validation.listing.reviewer.TestFunctionalityReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.UnattestedCriteriaWithDataReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.UnsupportedCharacterReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.UrlReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.ValidDataReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.edition2014.RequiredData2014Reviewer;
+import gov.healthit.chpl.validation.listing.reviewer.edition2014.TestFunctionality2014Reviewer;
 import gov.healthit.chpl.validation.listing.reviewer.legacy.LegacyTestToolReviewer;
 
 /**
@@ -64,8 +64,8 @@ public class Edition2014LegacyListingValidator extends Validator {
     private LegacyTestToolReviewer ttReviewer;
 
     @Autowired
-    @Qualifier("testFunctionalityReviewer")
-    private TestFunctionalityReviewer tfReviewer;
+    @Qualifier("testFunctionality2014Reviewer")
+    private TestFunctionality2014Reviewer tfReviewer;
 
     @Autowired
     @Qualifier("urlReviewer")
